feat(advance-runner)!: validate snapshot hash
The snapshot's template hash gets compared with its counterpart
in the blockchain, causing the advance-runner to stop if it does not
match.

BREAKING CHANGE: adds the PROVIDER_HTTP_ENDPOINT parameter, which is
required to validate snapshots. Validation is enabled by default, but
can be disabled by setting SNAPSHOT_VALIDATION_ENABLED to false