fix: service reachability
When executing the cartesi-rollups-node in a docker container, using the address 127.0.0.1 blocks services from receiving connections from interfaces outside the container. This prevents users from interacting with the Host Runner or querying the GraphQL and Inspect servers.

This commit also fixes the port mapping for GraphQL and Inspect servers in the node-compose.yml file