fix: panic when executable not in $PATH
The Service struct used to evaluate its termination reason to decide if it would return an error or not. This code didn't check for nil pointers and could panic during runtime.

Instead of evaluating the exit code or retrieving the received signal, we changed the design to always exit with an error. It now can be a `context.Canceled` or `exec.ExitError`.

The new code will ignore errors occurring during the exit process after context cancelation. This is by design since this is a rare edge case and the Supervisor is only interested in terminating the process.