feat: hide internal services behind proxy
Modifies HttpService to act as a reverse proxy. Calls to /graphql, /inspect, /rollups, and /metrics endpoints are now received by it and forwarded to graphql-server, inspect-server, host-runner, and dispatcher respectively. The ports used by these services needn't be exposed anymore, making HttpService the single point of access to the node's APIs.

This a is a step towards the goal of transforming these services into golang modules.