chore: add set -e for RUN heredoc instruction
When replacing RUN apt update && apt install ... we were making use of
&&, that ensures that the following command will only execute if the
previous one succeeds. When using the RUN <<EOF format, we lose this and
can have some susprises during the docker build process. It's common for
a command inside an RUN heredoc type to have a filing process but that
whole instruction pass as a success in the build, and the result will
only be noticed on the runtime image.

Adding set -e will ensure that if any command fails, the whole
instruction will fail.